Miglioramenti di gioco e commenti delle function

+++ header.h
@@ -12,8 +12,8 @@ int field[40] = {1,2,3,4,5,6,7,8,9,10,
 
 typedef struct giocatore{
     char *name;
-    int point;             //Struttura per la gestione dei giocatori,nome_giocatore,puntenggio_giocatore,avversario per lo switch.
-    int opponent;
+    unsigned int point;             //Struttura per la gestione dei giocatori,nome_giocatore,puntenggio_giocatore,avversario per lo switch.
+    unsigned short opponent;
 }player;
 
 player player1 = {"Giocatore1",0,2};
@@ -21,7 +21,7 @@ player player2 = {"Giocatore2",0,1};
 
 player *ptr;  // puntantore per la gestione dei player
 
-extern int ATT_PLAYER = 0;  // Variabile esterna per controllare il turno di ogni giocatore.
+extern unsigned short ATT_PLAYER = 0;  // Variabile esterna per controllare il turno di ogni giocatore.
 
 
 // Funzioni gestionali
